tts/requirements.txt
openai>=1.45.0
python-dotenv
